requirements.txt
os
streamlit
pypdf2
python-dotenv
langchain
langchain-community
langchain-groq
langchain-mistralai
faiss-cpu